+++ src/main/java/com/chungnamthon/cheonon/poweruser/service/PowerUserService.java
@@ -3,7 +3,7 @@ package com.chungnamthon.cheonon.poweruser.service;
 import com.chungnamthon.cheonon.poweruser.domain.PowerUser;
 import com.chungnamthon.cheonon.poweruser.dto.PowerUserResponse;
 import com.chungnamthon.cheonon.point.repository.PointRepository;
-import com.chungnamthon.cheonon.point.repository.PowerUserRepository;
+import com.chungnamthon.cheonon.poweruser.PowerUserRepository;
 import com.chungnamthon.cheonon.user.domain.User;
 import com.chungnamthon.cheonon.user.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
